backend/requirements.txt
# Minimal dependencies for the agentic backend

# Anthropic API
anthropic>=0.39.0

# Nano Banana (Gemini 2.5 Flash Image) for reference image generation
google-genai>=0.8.0

# Google Gemini API for image generation (Character Identity)
google-generativeai>=0.3.0

# Image processing
pillow>=10.0.0

# Environment variables
python-dotenv>=1.0.0

# API layer for Character Development System
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
